create readme
quick afternoon project playing around with breadth-first search algorithms, using tkinter for visualization

run `python main.py` at the terminal to pull up a very basic tkinter GUI

create the graph using at least 3 clicks:
- first click will place the start node (green)
- second click will place the end node (red)
- third click (and higher) will create obstacles.  Click and drag is supported here

Press the (barely visible?) button at the bottom of the GUI to begin the path finding algorithm.  Visited nodes will be shaded grey.  When the end node is reached, the shortest path identified will be highlighted in blue.  The length of this path will be output in the terminal.

To clear the board, just close the window and rerun the script.